requirements.txt
requests
tenacity
python-dotenv
beautifulsoup4
feedparser
regex
warcio
adaption
